Create new file for get categories endpoint
