+++ src/js/utils.js
@@ -1,7 +1,8 @@
 export const API = {
   base: 'https://www.themealdb.com/api/json/v1/1',
   search: '/search.php',
-  random: '/random.php'
+  random: '/random.php',
+  categories: '/categories.php'
 }
 
 export function getDataInJson(url) {
